third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/debugger-ui/function-details-expected.txt

Tests that Debugger.getFunctionDetails command returns correct location. Bug 71808


Running: testGetFirstLineFunctionDetails
firstLineFunction type = function
Function details: 
lineNumber: 9
columnNumber: 34
scriptId is valid: true
functionName: firstLineFunction
scopeChain #0: Global; <global object properties omitted>

Running: testGetNonFirstLineFunctionDetails
notFirstLineFunction type = function
Function details: 
lineNumber: 14
columnNumber: 37
scriptId is valid: true
functionName: notFirstLineFunction
scopeChain #0: Global; <global object properties omitted>

Running: testGetDetailsOfFunctionWithInferredName
obj.m type = function
Function details: 
lineNumber: 20
columnNumber: 23
scriptId is valid: true
functionName: m
scopeChain #0: Global; <global object properties omitted>

Running: testGetDetailsOfFunctionWithDisplayName
functionWithDisplayName type = function
Function details: 
lineNumber: 23
columnNumber: 40
scriptId is valid: true
functionName: user-friendly name
scopeChain #0: Global; <global object properties omitted>

Running: testGetDetailsOfFunctionWithDisplayNameGetter
functionWithDisplayNameGetter type = function
Function details: 
lineNumber: 26
columnNumber: 46
scriptId is valid: true
functionName: functionWithDisplayNameGetter
scopeChain #0: Global; <global object properties omitted>

Running: testSmallClosure
smallClosure type = function
Function details: 
lineNumber: 29
columnNumber: 57
scriptId is valid: true
functionName: 
scopeChain #0: Closure; p: "Capybara"
scopeChain #1: Global; <global object properties omitted>

Running: testBigClosure
bigClosure type = function
Function details: 
lineNumber: 42
columnNumber: 35
scriptId is valid: true
functionName: 
scopeChain #0: Catch; ex: <no string representation>
scopeChain #1: With Block; e: 7,f: 5,u: <no value>,v: <no value>
scopeChain #2: Closure; p: <no string representation>
scopeChain #3: Global; <global object properties omitted>

Running: testGenFunction
gen type = function
Function details: 
lineNumber: 49
columnNumber: 21
scriptId is valid: true
functionName: gen
isGenerator: true
scopeChain #0: Global; <global object properties omitted>
